backend/pyproject.toml
[project]
name = "bloom-art"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "commitizen>=4.9.1",
    "django>=5.2.7",
    "django-jazzmin>=3.0.1",
    "django-jet>=1.0.8",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.11",
    "pillow>=12.0.0",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "ruff>=0.14.3",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true


[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle (PEP 8 style issues like spacing, indentation)
    "F",  # pyflakes (undefined variables, unused imports, etc.)
    "B",  # flake8-bugbear (likely bugs, dangerous patterns)
    "W",  # pycodestyle warnings (minor style warnings)
    "C",  # mccabe (complexity checker — too complex functions)
    "N",  # pep8-naming (enforces correct naming conventions)
    "A",  # flake8-builtins (avoid shadowing Python built-ins like `list`, `id`, etc.)
    "PT", # pytest-style – checks specific to pytest
    "I"  # isort – import sorting
]
ignore = ["E501"] # ignore line-too-long, if you want
